work commit

+++ src/main/java/com/example/jikedemo/dao/Product.java
@@ -3,19 +3,20 @@ package com.example.jikedemo.dao;
 import com.sun.xml.internal.ws.developer.Serialization;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class Product implements Serializable{
 
     private static final long serialVersionUID = 748392348L;
     private String name;
     private String description;
-    private float price;
+    private BigDecimal price;
 
     public Product() {
 
     }
 
-    public Product(String name, String description, float price) {
+    public Product(String name, String description, BigDecimal price) {
         this.name = name;
         this.description = description;
         this.price = price;
@@ -37,11 +38,11 @@ public class Product implements Serializable{
         this.description = description;
     }
 
-    public float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
